requirements.txt
fastapi==0.95.1
pyyaml
pydantic==1.10.7
pydantic_yaml==0.11.2
uvicorn==0.21.1
psycopg2-binary
